Refactor rabbit test queue route
