+++ backend/src/modules/rabbit/controllers/index.ts
@@ -1,2 +1,1 @@
-export * from "./addToTestQueue";
 export * from "./health";
