+++ backend/src/modules/rabbit/controllers/queue/test/index.ts
@@ -1,0 +1,1 @@
+export * from "./add";
