+++ backend/src/modules/rabbit/router.ts
@@ -1,9 +1,10 @@
 import { Router } from "express";
-import { addToTestQueue, health } from "./controllers";
+import { health } from "./controllers";
+import { add as addToTestQueue } from "./controllers/queue/test";
 
 const router = Router();
 
 router.get("/health", health);
-router.post("/addToTestQueue", addToTestQueue);
+router.post("/queue/test", addToTestQueue);
 
 export default router;
