+++ backend/src/modules/rabbit/utils/index.ts
@@ -1,2 +1,2 @@
 export * from "./connection";
-export * from "./queue";
+export * from "./testQueue";
