+++ backend/tests/modules/rabbit/router.test.ts
@@ -2,12 +2,16 @@ import type { Request, Response } from "express";
 import express from "express";
 import request from "supertest";
 import router from "@/modules/rabbit/router";
-import { addToTestQueue, health } from "@/modules/rabbit/controllers";
+import { health } from "@/modules/rabbit/controllers";
+import { add as addToTestQueue } from "@/modules/rabbit/controllers/queue/test";
 
 // Mock the controller functions
 jest.mock("@/modules/rabbit/controllers", () => ({
 	health: jest.fn((req: Request, res: Response) => res.status(200).send("Complete")),
-	addToTestQueue: jest.fn((req: Request, res: Response) => res.status(200).send("Complete")),
+}));
+
+jest.mock("@/modules/rabbit/controllers/queue/test", () => ({
+	add: jest.fn((req: Request, res: Response) => res.status(200).send("Complete")),
 }));
 
 // Helper function to create an Express app with the router
@@ -25,10 +29,10 @@ describe("Rabbit Router", () => {
 		expect(health).toHaveBeenCalled();
 	});
 
-	// Test case for /addToTestQueue route
-	it("should call the addToTestQueue controller on /addToTestQueue route", async () => {
+	// Test case for /queue/test route
+	it("should call the addToTestQueue controller on /queue/test route", async () => {
 		const app = createApp();
-		await request(app).post("/addToTestQueue").expect(200, "Complete");
+		await request(app).post("/queue/test").expect(200, "Complete");
 		expect(addToTestQueue).toHaveBeenCalled();
 	});
 });
